commit 18 Jun 2025 MoreScreen.kt and FeedsScreen.kt implemented

+++ app/src/main/java/com/example/hrmanagement/ui/leave/ApplyLeaveViewModel.kt
@@ -149,7 +149,7 @@ class ApplyLeaveViewModel(
             val leaveData = LeaveData(
                 lastLeaveId!!,
                 leaveTypeSelected.value,
-                differenceInDays,
+                differenceInDays.toFloat(),
                 fromDate.value,
                 toDate.value,
                 "Approved",
